Style hexo-renderer-markdown-it's automatic headline ID's
See https://github.com/celsomiranda/hexo-renderer-markdown-it/wiki/Advanced-Configuration#automatic-headline-ids

hexo-renderer-markdown can add permalinks to headings. However, they are
prepended to the heading and styled like a regular link, which doesn't look
that good. This commit moves it to the right, hides it when not hovering the
heading, and tones it down until the actual anchor is hovered.